refactor + cleanup
The project was clean removing non necessary files and ignoring some
that are not necessary.
The npm package was also created, to track project dependencies and
script.
Some files were renamed to improve readability.
The project was also upgrade to ecmascript 2015 syntax.
A README file was added to the project to keep some documentation.
And tests were refactored in general.